+++ Sources/RunnerLib/Files Import/getDangerJSPath.swift
@@ -1,5 +1,6 @@
 import Foundation
 import ShellOut
+import Logger
 
 public func getDangerCommandPath(_ command: String, logger: Logger) throws -> String {
     logger.debug("Finding out where the danger executable is")